refactoring of DICOM part
